refactor(utils): split utils
