+++ packages/csso/src/index.ts
@@ -8,7 +8,7 @@
  * Module dependencies.
  */
 import type { MinifierOptions } from "@node-minify/types";
-import { utils } from "@node-minify/utils";
+import { writeFile } from "@node-minify/utils";
 import { minify } from "csso";
 
 /**
@@ -23,7 +23,7 @@ export function csso({ settings, content, callback, index }: MinifierOptions) {
     const contentMinified = minify(content ?? "", settings?.options);
     if (settings && !settings.content && settings.output) {
         settings.output &&
-            utils.writeFile({
+            writeFile({
                 file: settings.output,
                 content: contentMinified.css,
                 index,
