+++ packages/jsonminify/src/index.ts
@@ -8,7 +8,7 @@
  * Module dependencies.
  */
 import type { MinifierOptions } from "@node-minify/types";
-import { utils } from "@node-minify/utils";
+import { writeFile } from "@node-minify/utils";
 import jsonminify from "jsonminify";
 
 /**
@@ -28,7 +28,7 @@ export function jsonMinify({
     const contentMinified = jsonminify(content ?? "");
     if (settings && !settings.content && settings.output) {
         settings.output &&
-            utils.writeFile({
+            writeFile({
                 file: settings.output,
                 content: contentMinified,
                 index,
