+++ packages/sqwish/src/index.ts
@@ -8,7 +8,7 @@
  * Module dependencies.
  */
 import type { MinifierOptions } from "@node-minify/types";
-import { utils } from "@node-minify/utils";
+import { writeFile } from "@node-minify/utils";
 import minify from "sqwish";
 
 /**
@@ -27,7 +27,7 @@ export function sqwish({
 }: MinifierOptions) {
     const contentMinified = minify.minify(content, settings?.options?.strict);
     if (settings && !settings.content && settings.output) {
-        utils.writeFile({
+        writeFile({
             file: settings.output,
             content: contentMinified,
             index,
