+++ packages/utils/__tests__/utils.test.ts
@@ -5,22 +5,28 @@
  */
 
 import { describe, expect, test } from "vitest";
-import { utils } from "../src/index.ts";
+import {
+    buildArgs,
+    getFilesizeGzippedInBytes,
+    getFilesizeInBytes,
+    prettyBytes,
+    readFile,
+    setFileNameMin,
+    writeFile,
+} from "../src/index.ts";
 
 const fixtureFile = `${__dirname}/../../../tests/fixtures/fixture-content.js`;
 
 describe("Package: utils", () => {
     describe("readFile", () => {
         test("should return the content", () =>
-            expect(utils.readFile(fixtureFile)).toMatch(
-                "console.log('content');"
-            ));
+            expect(readFile(fixtureFile)).toMatch("console.log('content');"));
     });
 
     describe("writeFile", () => {
         test("should return the content", () =>
             expect(
-                utils.writeFile({
+                writeFile({
                     file: `${__dirname}/../../../tests/tmp/temp.js`,
                     content: "const foo = 'bar';",
                 })
@@ -30,7 +36,7 @@ describe("Package: utils", () => {
     describe("buildArgs", () => {
         test("should return an array with args", () =>
             expect(
-                utils.buildArgs({
+                buildArgs({
                     foo: "bar",
                 })
             ).toEqual(["--foo", "bar"]));
@@ -38,15 +44,13 @@ describe("Package: utils", () => {
 
     describe("getFilesizeInBytes", () => {
         test("should return file size", () =>
-            expect(utils.getFilesizeInBytes(fixtureFile)).toMatch(
-                /(24 B)|(25 B)/
-            ));
+            expect(getFilesizeInBytes(fixtureFile)).toMatch(/(24 B)|(25 B)/));
     });
 
     describe("getFilesizeGzippedInBytes", () => {
         test("should return file size", (): Promise<void> =>
             new Promise<void>((done) => {
-                utils.getFilesizeGzippedInBytes(fixtureFile).then((size) => {
+                getFilesizeGzippedInBytes(fixtureFile).then((size) => {
                     expect(size).toMatch(/(44 B)|(45 B)/);
                     done();
                 });
@@ -56,19 +60,17 @@ describe("Package: utils", () => {
     describe("pretty bytes", () => {
         test("should throw when not a number", () => {
             // @ts-expect-error
-            expect(() => utils.prettyBytes("a")).toThrow();
+            expect(() => prettyBytes("a")).toThrow();
         });
 
         test("should return a negative number", () =>
-            expect(utils.prettyBytes(-1)).toBe("-1 B"));
+            expect(prettyBytes(-1)).toBe("-1 B"));
 
-        test("should return 0", () => expect(utils.prettyBytes(0)).toBe("0 B"));
+        test("should return 0", () => expect(prettyBytes(0)).toBe("0 B"));
     });
 
     describe("setFileNameMin", () => {
         test("should return file name min", () =>
-            expect(utils.setFileNameMin("foo.js", "$1.min.js")).toBe(
-                "foo.min.js"
-            ));
+            expect(setFileNameMin("foo.js", "$1.min.js")).toBe("foo.min.js"));
     });
 });
