+++ packages/utils/src/prettyBytes.ts
@@ -1,0 +1,39 @@
+/*!
+ * node-minify
+ * Copyright(c) 2011-2024 Rodolphe Stoclin
+ * MIT Licensed
+ */
+
+/**
+ * Get the size in human readable.
+ * From https://github.com/sindresorhus/pretty-bytes
+ * @param num Number
+ */
+export function prettyBytes(num: number): string {
+    const UNITS = ["B", "kB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"];
+
+    if (!Number.isFinite(num)) {
+        throw new TypeError(
+            `Expected a finite number, got ${typeof num}: ${num}`
+        );
+    }
+
+    const neg = num < 0;
+
+    if (neg) {
+        num = -num;
+    }
+
+    if (num < 1) {
+        return `${(neg ? "-" : "") + num} B`;
+    }
+
+    const exponent = Math.min(
+        Math.floor(Math.log(num) / Math.log(1000)),
+        UNITS.length - 1
+    );
+    const numStr = Number((num / 1000 ** exponent).toPrecision(3));
+    const unit = UNITS[exponent];
+
+    return `${(neg ? "-" : "") + numStr} ${unit}`;
+}
